requirements.txt
altgraph==0.17.4
apturl==0.5.2
blinker==1.4
Brlapi==0.7.0
cachetools==5.5.2
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
contourpy==1.1.1
crypto==1.4.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
entrypoints==0.3
fido2==0.8.1
fonttools==4.57.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==2.8
importlib-metadata==8.5.0
importlib-resources==6.4.5
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
matplotlib==3.7.5
Naked==0.1.32
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
packaging==25.0
pandas==2.0.3
pexpect==4.6.0
Pillow==7.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.4
PyQt5==5.14.1
pyRFC3339==1.1
pyscard==1.9.9
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
pytz==2025.2
pyusb==1.0.2
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
rsa==4.9.1
scipy==1.10.1
SecretStorage==2.3.1
shellescape==3.8.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
systemd-python==234
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
wadllib==1.3.3
xkit==0.0.0
yubikey-manager==3.1.1
zipp==3.20.2
